- Move the java code not related specifically to src/main/java - Create new package names to better organize the code - Migrate to jackson3
