+++ tests/pom.xml
@@ -14,6 +14,8 @@
         <maven.compiler.source>21</maven.compiler.source>
         <maven.compiler.target>21</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+
+        <jackson-3.version>3.2.2</jackson-3.version>
         <junit.version>5.12.2</junit.version>
         <surefire.version>3.5.2</surefire.version>
 
@@ -41,16 +43,14 @@
             <scope>test</scope>
         </dependency>
         <dependency>
-            <groupId>com.fasterxml.jackson.core</groupId>
+            <groupId>tools.jackson.core</groupId>
             <artifactId>jackson-databind</artifactId>
-            <version>2.18.3</version>
-            <scope>test</scope>
+            <version>${jackson-3.version}</version>
         </dependency>
         <dependency>
-            <groupId>com.fasterxml.jackson.dataformat</groupId>
+            <groupId>tools.jackson.dataformat</groupId>
             <artifactId>jackson-dataformat-yaml</artifactId>
-            <version>2.18.3</version>
-            <scope>test</scope>
+            <version>${jackson-3.version}</version>
         </dependency>
     </dependencies>
