+++ tests/src/main/java/io/quarkus/ai/runner/opencode/OpenCodeSessionExporter.java
@@ -1,7 +1,7 @@
-package io.quarkus.migration.runner;
+package io.quarkus.ai.runner.opencode;
 
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
+import tools.jackson.core.type.TypeReference;
+import tools.jackson.databind.ObjectMapper;
 
 import java.io.BufferedReader;
 import java.io.File;
